+++ product_pack/__init__.py
@@ -3,6 +3,7 @@
 # For copyright and license notices, see __openerp__.py file in root directory
 ##############################################################################
 from . import pack
+from . import product
 from . import sale_order_line_pack_line
 from . import sale_order_line
 from . import sale_order
